% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addins}
\alias{addins}
\alias{insertImgCenterAddin}
\alias{insertSpanAddin}
\alias{insertLaTeXAddin}
\alias{insertRoxyLinkAddin}
\title{Rstudio addins}
\usage{
insertImgCenterAddin()

insertSpanAddin()

insertLaTeXAddin()

insertRoxyLinkAddin()
}
\value{
<img src="" style="display:block; margin: 0 auto;"> as text.

<span class=""></span> as text.
}
\description{
Rstudio addins, mostly for Rmarkdown html/css snippets
}
\details{
- insertImgCenter This function is an Rstudio addin that inserts <img src="" style="display:block; margin: 0 auto;"> so that one can have a centered image on the fly without an R chunk. You must supply the image location for src.

- insertSpan Insert a span for a particular class.
}
\examples{
\dontrun{
library(lazerhawk)
insertImgCenterAddin()
insertSpanAddin()
}


}
