% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_SummaryTable.R
\name{brms_SummaryTable}
\alias{brms_SummaryTable}
\title{Summarize brms-fit objects}
\usage{
brms_SummaryTable(model, formatOptions = list(digits = 2, nsmall = 2),
  round = 2, astrology = F, hype = F, panderize = F, justify = NULL,
  ...)
}
\arguments{
\item{model}{A brmsfit object}

\item{formatOptions}{A list(!) of options to pass to the format function}

\item{round}{See code{\link[base]{round}}}

\item{astrology}{Put asterisks next to effects whose credible interval does not
contain 0}

\item{hype}{Run one-sided hypothesis tests of whether effect is greater
(less) than 0 if positive (negative)}

\item{panderize}{Create a markdown summary table. Requires the pander package}

\item{justify}{Additional argument to pass to \code{\link[pander]{pander}}}

\item{...}{Additional arguments to pass to \code{\link[brms]{summary.brmsfit}}}
}
\value{
A data.frame, or Pandoc markdown table if panderize=T.
}
\description{
Returns a summary table of model output from the brms package.
  Requires the brms package.
}
\details{
This function creates a data.frame summary object for a brms package
  model object.  As of now it only does so for the fixed effects part of the
  model.  It will star 'significant' effects, add results from one-sided
  hypothesis tests, and allow additional formating options.

  For \code{hype = TRUE}, the returned object contains a column indicating
  the posterior probability under the hypothesis against its alternative as
  noted in the hypothesis function of the brms package. Both the evidence
  ratio and probability are produced. Note that for effects near zero, the
  probability can be < .5 due to the posterior samples not being perfectly
  symmetric.
}
\examples{

\dontrun{

## Probit regression using the binomial family
library(brms)
n <- sample(1:10, 100, TRUE)  # number of trials
success <- rbinom(100, size = n, prob = 0.4)
x <- rnorm(100)
d <- data.frame(n, success, x)
fit4 <- brm(success | trials(n) ~ x, data=d, family = binomial("probit"))
summary(fit4)

library(lazerhawk)
brms_SummaryTable(fit4)
brms_SummaryTable(fit4, astrology=TRUE, hype=TRUE)
brms_SummaryTable(fit4, astrology=TRUE, hype=TRUE, panderize=TRUE, justify='lrrrrclr')
}

}
\seealso{
\code{\link[brms]{summary.brmsfit}} \code{\link[brms]{hypothesis}}
  \code{\link[rstan]{print.stanfit}} \code{\link[base]{format}}
  \code{\link[base]{round}}
}
