% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrheat.R
\name{corrheatOutput}
\alias{corrheatOutput}
\alias{renderCorrheat}
\title{Wrapper functions for using corrheat in shiny}
\usage{
corrheatOutput(outputId, width = "100\%", height = "400px")

renderCorrheat(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width, height}{The width and height of the map (see
\link[htmlwidgets]{shinyWidgetOutput})}

\item{expr}{An expression that generates a \code{\link[d3heatmap]{d3heatmap}} object}

\item{env}{The environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Use \code{corrheatOutput} to create a UI element, and \code{renderCorrheat}
to render the heatmap.
}
\details{
Note that this is not a priority in the least for me at this time,
  but I went ahead and ported the d3heatmap code and at least the following
  example will work fine.
}
\examples{
\donttest{
library(lazerhawk)
library(shiny)

ui <- fluidPage(
  h1("A corrheat demo"),
  selectInput("palette", "Palette", c("YlOrRd", "RdYlBu", "Greens", "Blues")),
  corrheatOutput("heatmap")
)

server <- function(input, output, session) {
  output$heatmap <- renderCorrheat({
    corrheat(cor(mtcars), colors = input$palette)
  })
}

shinyApp(ui, server)
}

}
\seealso{
\code{\link{corrheat}}, \code{\link[d3heatmap]{d3heatmapOutput}},
\code{\link[d3heatmap]{renderD3heatmap}}
}
