% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corr.R
\name{create_corr}
\alias{create_corr}
\title{Create a covariance/correlation matrix}
\usage{
create_corr(x, diagonal = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{diagonal}{A vector of variances}
}
\description{
Creates a correlation or covariance matrix given a vector of
  values for the lower triangle.
}
\details{
This function takes the provided values and creates a symmetric (covariance) matrix, e.g. for simulation or demonstration purposes.  The values are filled column-wise in the lower triangle first. For example \code{mat[1,2] = x[1]}, \code{mat[1,3] = x[2]} and so on.
}
\examples{
library(lazerhawk)
cors = runif(6, -1, 1)
create_corr(cors)
covs = runif(3, 1, 5)
create_corr(covs, diag=c(1,2,3))
}
