% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insertImgCenterAddin}
\alias{insertImgCenterAddin}
\title{Return an img tag for a centered image}
\usage{
insertImgCenterAddin()
}
\value{
<img src="" style="display:block; margin: 0 auto;"> as text.
}
\description{
Inserts html for a centered image
}
\details{
This function is an Rstudio addin that inserts <img src="" style="display:block; margin: 0 auto;"> so that one can have a centered image on the fly.
}
\examples{
## Not Run
# library(lazerhawk)
# insertImgCenterAddin()

}

