% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_summary.R
\name{num_summary}
\alias{num_summary}
\title{num_summary}
\usage{
num_summary(x, digits = 1)
}
\arguments{
\item{x}{A numeric or logical variable}

\item{digits}{Argument to \code{round}. Default is 1.}
}
\value{
A data.frame containing those values.
}
\description{
Essentially a slight tweak to the base R summary.
}
\details{
Takes a numeric variable and provides sample size, mean, standard deviation, min, first quartile,
  median, third quartile, and max, and number of NAs.
}
\examples{
num_summary(c(1:10, NA))
num_summary(c('1','2','3'))
}
