% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise}
\alias{pairwise}
\title{Pairwise operations}
\usage{
pairwise(x, margin = 2, fun)
}
\arguments{
\item{x}{A matrix class object or data.frame that can be coerced to a numeric matrix.}

\item{margin}{Work on rows (1) or columns (2)}

\item{fun}{The function to apply. Must be applicable to numeric data.}
}
\description{
Apply a function to all possible pairwise combinations of rows or columns.
}
\details{
The function must work on all rows (columns) provided.  While a
  data.frame is accepted, it will be coerced to a matrix, and so its rows
  (columns) must be atomic vectors that are not character strings.  While it
  might be expanded to be more flexible at some point, it should be thought
  of as a way to create things like distance or covariance matrices with
  arbitrary functions. At some point a parallel version will be added.
}
\examples{
library(lazerhawk)
# dot product
pairwise(mtcars, 1,  function(x,y) sum(x*y))

# correlation
pairwise(mtcars, 2,  cor)

# correlation for factors
mydata = data.frame(a = factor(sample(letters[1:3], 10, replace=TRUE)),
                    b = factor(sample(letters[1:4], 10, replace=TRUE)))
pairwise(mydata, 2,  cor)
}

