% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_NA.R
\name{sum_NA}
\alias{sum_NA}
\alias{sum_NaN}
\alias{sum_blank}
\title{Investigate missingness}
\usage{
sum_NA(x)

sum_NaN(x)

sum_blank(x)
}
\arguments{
\item{x}{a vector with multiple elements}
}
\value{
The number of non-elements so defined
}
\description{
Investigate missingness
}
\details{
These functions return the number of NA, NaN, or blank values in a
  vector.  Such information can serve as a diagnostic or used outright, e.g.
  in longitudinal settings. If you are dealing with singletons, you should
  just use the underlying function, e.g. is.na. For total blank values, the
  stringi package is required.
}
\examples{
library(lazerhawk)
sum_blank(c('    ', '', 'c', 'd'))
sum_NA(c(NA, '', 'c', NA))
sum_NaN(c(1:3, NA, NaN))
sum_NA(1:4)



}
