% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggClean.R
\name{theme_trueMinimal}
\alias{theme_trueMinimal}
\alias{theme_plotly}
\title{Clean visualizations}
\usage{
theme_trueMinimal()

theme_plotly(vis)
}
\arguments{
\item{vis}{A plotly, or ggvis}
}
\description{
Clean up plots from their defaults.
}
\details{
From a gray background, to unnecessary gridlines, to by-default reference lines, some of the more popular visualization packages come out 75% fantastic and 25% wtf?.  These functions remove unnecessary gridlines, 'de-bold' the blacks, etc.
**ggplot2**: This function takes a ggplot object and removes the gray background, gridlines and adds opacity to the default black axes and labels, allowing the pattern of the visual to be expressed in unimpeded fashion.

Plotly will put reference lines at zero, and some of its 'modebar' is unnecessary.  Otherwise little is changed at this point.
}
\examples{
require(lazerhawk)
require(ggplot2)
require(plotly)
data(mtcars)
mtcars \%>\% ggplot(aes(wt, mpg)) + theme_trueMinimal()
mtcars \%>\% plot_ly(x=~wt, y=~mpg, mode='markers') \%>\% theme_plotly()
}
