% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github_pkgs.R
\name{update_github_pkgs}
\alias{update_github_pkgs}
\title{Update GitHub packages.}
\usage{
update_github_pkgs()
}
\description{
Updates packages installed from GitHub.
}
\details{
Updates packages installed from GitHub. Rather, it re-installs
  GitHub-only packages.  This is not quite as smooth a process as with CRAN,
  but you'll get a note of \code{FALSE} for the package if it failed.

Taken entirely from user
  vh-d's response on
  \href{https://stackoverflow.com/questions/32538052/update-all-packages-from-github}{StackOverflow}
  with only a very slight modification.
}
\examples{
\dontrun{
update_github_pkgs()
}
}
