% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrheat.R
\name{corrheat}
\alias{corrheat}
\title{D3 Heatmap widget for correlation matrices}
\source{
Base code comes from \link[d3heatmap]{d3heatmap} package's core function,
which was mostly gutted, and all dendrogram functionality replaced with a
factor analytic approach.
}
\usage{
corrheat(x, rotation = "none", ordering = "first", labRow = rownames(x),
  labCol = colnames(x), cexRow, cexCol, digits = 3L, cellnote,
  cellnote_scale = FALSE, theme = NULL, colors = "RdBu", width = NULL,
  height = NULL, xaxis_height = 80, yaxis_width = 120,
  xaxis_font_size = NULL, yaxis_font_size = NULL, brush_color = "#0000FF",
  show_grid = TRUE, anim_duration = 500, ...)
}
\arguments{
\item{x}{A complete correlation matrix}

\item{rotation}{Preferred rotation for factor analysis solution (see
\link[stats]{factanal})}

\item{ordering}{Order cols/rows based on max raw ("raw") or absolute
("absolute") loadings across all factors. Otherwise just based on first
factor's raw loadings ("first" default).}

\item{labRow}{character vectors with row labels to use (from top to bottom);
default to rownames(x).}

\item{labCol}{character vectors with column labels to use (from left to
right); default to colnames(x).}

\item{cexRow}{positive numbers. If not missing, it will override
\code{xaxis_font_size} and will give it a value cexRow*14}

\item{cexCol}{positive numbers. If not missing, it will override
\code{yaxis_font_size} and will give it a value cexCol*14}

\item{digits}{integer indicating the number of decimal places to be used by
\link{round} for 'label'.}

\item{cellnote}{(optional) matrix of the same dimensions as \code{x} that has
the human-readable version of each value, for displaying to the user on
hover. If \code{NULL}, then \code{x} will be coerced using
\code{\link{as.character}}. If missing, it will use \code{x}, after
rounding it based on the \code{digits} parameter.}

\item{cellnote_scale}{logical (default is FALSE). IF cellnote is missing and
x is used, should cellnote be scaled if x is also scaled?}

\item{theme}{A custom CSS theme to use. Currently the only valid values are
\code{""} and \code{"dark"}. \code{"dark"} is primarily intended for
standalone visualizations, not R Markdown or Shiny.}

\item{colors}{Either a colorbrewer2.org palette name (e.g. \code{"YlOrRd"} or
\code{"Blues"}), or a vector of colors to interpolate in hexadecimal
\code{"#RRGGBB"} format, or a color interpolation function like
\code{\link[grDevices]{colorRamp}}.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{xaxis_height}{Size of axes, in pixels.}

\item{yaxis_width}{Size of axes, in pixels.}

\item{xaxis_font_size}{Font size of axis labels, as a CSS size (e.g. "14px"
or "12pt").}

\item{yaxis_font_size}{Font size of axis labels, as a CSS size (e.g. "14px"
or "12pt").}

\item{brush_color}{The base color to be used for the brush. The brush will be
filled with a low-opacity version of this color. \code{"#RRGGBB"} format
expected.}

\item{show_grid}{\code{TRUE} to show gridlines, \code{FALSE} to hide them, or
a numeric value to specify the gridline thickness in pixels (can be a
non-integer).}

\item{anim_duration}{Number of milliseconds to animate zooming in and out.
For large \code{x} it may help performance to set this value to \code{0}.}

\item{...}{currently ignored}
}
\description{
Creates a D3.js-based heatmap widget specifically geared toward
  correlation matrices. All dendrogram related functionality related to the
  original \code{d3heatmap} function is removed, and replaced with an
  ordering based on a factor analysis.
}
\details{
d3heatmap is a great tool for matrix visualization and highly recommended.
Correlation matrices are typically better visualized rather than parsed
numerically, and while one can do so with d3heatmap, one often may not want
the cluster based approach to ordering if only dealing with a correlation
matrix, which are often too small to be useful for a cluster based approach.
This function does not currently allow choice of the number of factors
(\code{nfact = 1} if ncol <=4, else \code{floor(sqrt(ncol(x)))}). If you want
explore a factor analysis you will have to do that separately. In addition,
the choice of color will be overridden if the correlations are mostly/all
positive or negative. Right now this function uses \code{factanal} for the
underlying factor analysis, but in the future will use psych's \code{fa}
function (psych has a corrplot function also).
}
\examples{
library(lazerhawk)
corrheat(cor(mtcars), rotation='varimax')
corrheat(Harman74.cor$cov)

}
\seealso{
\code{\link[psych]{fa}}, \code{\link[stats]{factanal}}
}

