% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_all.R
\name{describe_all}
\alias{describe_all}
\alias{describe_all_num}
\alias{describe_all_cat}
\alias{describeAll}
\title{describe_all Describe your data cleanly and effectively}
\usage{
describe_all(data, digits = 2, NAcat_include = TRUE)

describe_all_num(data, digits = 2)

describe_all_cat(data, digits = 2, NAcat_include = TRUE)

describeAll(data, digits = 2, NAcat_include = TRUE)
}
\arguments{
\item{data}{The dataset, of class data.frame.}

\item{digits}{See \code{\link[base]{round}}.}

\item{NAcat_include}{Include NA values as categorical levels? Default is
TRUE.}
}
\value{
A list with two elements of summaries for numeric and other variables
  respectively.
}
\description{
Describe data sets with multiple variable types effectively.
}
\details{
This function comes out of my frustrations from various data set
  summaries either being inadequate for my needs, too 'busy' with output or
  unable to deal well with mixed data types. Numeric data is treated entirely
  separately, and provides the same information as in num_by.  Categorical
  variables are summarized with frequencies and percentages. For empty
  categorical variables (e.g. after a subset), a warning is thrown.

  The functions \code{describe_all_num} and  \code{describe_all_cat} will
  provide only numeric or only categorical data summaries respectively.
  \code{describeAll} is a deprecated alias.
}
\examples{
library(lazerhawk); library(dplyr)
X = data.frame(f1 =gl(2, 1, 20, labels=c('A', 'B')), f2=gl(2, 2, 20, labels=c('X', 'Q')))
X = X \%>\% mutate(bin1 = rbinom(20, 1, p=.5),
                 logic1 = sample(c(TRUE, FALSE), 20, replace = TRUE),
                 num1 = rnorm(20),
                 num2 = rpois(20, 5),
                 char1 = sample(letters, 20, replace = TRUE))
describeAll(X)

describeAll(data.frame(x=factor(1:7)), digits=5)
}
\seealso{
\code{\link[base]{summary}} \code{\link[lazerhawk]{num_by}}
  \code{\link[lazerhawk]{num_summary}}
}
