% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeAll.R
\name{describeAll}
\alias{describeAll}
\title{Describe your data cleanly and effectively}
\usage{
describeAll(data, digits = 2)
}
\arguments{
\item{data}{The dataset, of class data.frame.}

\item{digits}{See \code{\link[base]{round}}.}
}
\description{
Describe data sets with multiple variable types effectively.
}
\details{
This function comes out of my frustrations from various data set
  summaries either being inadequate for my needs, too 'busy' with output or
  unable to deal well with mixed data types. Numeric data is treated entirely
  separately, and provides the same information as in num_by.  Categorical
  variables are summarized with frequencies and percentages.
}
\examples{
library(lazerhawk); library(dplyr)
X = data.frame(f1 =gl(2, 1, 20, labels=c('A', 'B')), f2=gl(2, 2, 20, labels=c('X', 'Q')))
X = X \%>\% mutate(bin1 = rbinom(20, 1, p=.5),
                 logic1 = sample(c(TRUE, FALSE), 20, replace = TRUE),
                 num1 = rnorm(20),
                 num2 = rpois(20, 5),
                 char1 = sample(letters, 20, replace = TRUE))
describeAll(X)

describeAll(data.frame(x=factor(1:7)), digits=5)
}
\seealso{
\code{\link[base]{summary}} \code{\link[lazerhawk]{num_by}}
}
