% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_by.R
\name{num_by}
\alias{num_by}
\title{Summarize data}
\usage{
num_by(df, main_var, group_var)
}
\arguments{
\item{df}{data frame}

\item{main_var}{the main numeric variable to be summarised}

\item{group_var}{the optional grouping variable}
}
\value{
data.frame/tibble with the corresponding summary statistics
}
\description{
Provide common numeric summary information.
}
\details{
This function takes a numeric variable from a dataframe and provides
  sample size, mean, standard deviation, min, first quartile, median, third
  quartile, and max, possibly over a grouping variable. It works in the dplyr
  style using unquoted (bare) variable names.
}
\examples{
library(dplyr)
df1 <- tibble(
g1 = sample(1:2, 50, replace = TRUE),
g2 = sample(1:4, 50, replace = TRUE),
a = rnorm(50),
b = rpois(50, 10)
)


num_by(df1, main_var = a)
num_by(df1, main_var = a, group_var = g2)

num_by(df1, main_var = b)
num_by(df1, main_var = b, group_var = g1)


}
\seealso{
describeAll
}
